refactor: move log methods
